WIP: centralize edit tests in useQueryMutation; tests pass with console errors
