+++ client/src/Pages/About/ReviewQuotes/Tests/ReviewQuotes-display.test.tsx
@@ -2,15 +2,17 @@
 import { fireEvent } from '@testing-library/react';
 import App from 'Pages/App/App';
 import React from 'react';
-import { renderWithRouterAndProvider } from 'TestUtils/renderWith';
+import {
+  renderWithRouterAndProvider,
+  renderWithRouterProviderAndUser,
+} from 'TestUtils/renderWith';
 
 test('Renders four ReviewQuotes and course links for non-error server response', async () => {
   // Note: mocked server response is handled by msw, in the src/mocks folder
   // and src/setupTests.js. The handler is set to return
   // TestUtils/Data/testReviewQuotesData (see above) for /api/ReviewQuotes
 
-  // render entire App so that we can check Loading and Error
-  const screen = renderWithRouterAndProvider(<App />);
+  const { screen } = renderWithRouterAndProvider(<App />);
 
   // click the 'about' tab to trigger the ReviewQuotes retrieval
   const aboutNavLink = screen.getByRole('tab', { name: /about/ });
@@ -24,4 +26,16 @@ test('Renders four ReviewQuotes and course links for non-error server response',
   const courseLinks = screen.getAllByRole('link', { name: /Course \d/ });
   expect(courseLinks.length).toBe(5);
   courseLinks.forEach((course) => expect(course).toHaveAttribute('href'));
+});
+
+test('Renders quotes when logged in', async () => {
+  const loadingScreen = await renderWithRouterProviderAndUser(<App />);
+
+  // click the 'about' tab to trigger the review talks retrieval
+  const aboutLink = loadingScreen.getByRole('tab', { name: /about/ });
+  fireEvent.click(aboutLink);
+
+  // check quotes
+  const quotes = await loadingScreen.findAllByText(/body \d/);
+  expect(quotes.length).toBe(5);
 });
