+++ client/src/Pages/App/Tests/App.test.tsx
@@ -5,7 +5,7 @@ import App from '../App';
 
 describe('App rendering tests', () => {
   test('it starts at the home page', () => {
-    const screen = renderWithRouterAndProvider(<App />);
+    const { screen } = renderWithRouterAndProvider(<App />);
     const nameHeading = screen.getByRole('heading', {
       name: 'Bonnie Schulkin',
     });
@@ -13,13 +13,13 @@ describe('App rendering tests', () => {
   });
 
   test('page contains nav tabs', () => {
-    const screen = renderWithRouterAndProvider(<App />);
+    const { screen } = renderWithRouterAndProvider(<App />);
     const tabList = screen.getByRole('tablist', { name: '' });
     expect(tabList).toBeInTheDocument();
   });
 
   test('loading overlay is not visible', () => {
-    const screen = renderWithRouterAndProvider(<App />);
+    const { screen } = renderWithRouterAndProvider(<App />);
     const loadingSpinner = screen.queryByRole('progressbar', {
       hidden: true,
     });
@@ -27,7 +27,7 @@ describe('App rendering tests', () => {
   });
 
   test('error "snack" is not visible', () => {
-    const screen = renderWithRouterAndProvider(<App />);
+    const { screen } = renderWithRouterAndProvider(<App />);
     const errorAlert = screen.queryByRole('alert');
     expect(errorAlert).not.toBeInTheDocument();
   });
